% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Slurm_scripts.R
\name{Create_Slurm_scripts}
\alias{Create_Slurm_scripts}
\title{Create a Slurm parallel job to download ERA5}
\usage{
Create_Slurm_scripts(py.PATH, mem_per_cpu, time_limit, job_name, email)
}
\arguments{
\item{py.PATH}{Path to the created Python scripts}

\item{mem_per_cpu}{Memory per CPU in MB.}

\item{time_limit}{Time limit for the complete Slurm job}

\item{job_name}{The name of the job}

\item{email}{The user email as \email{"user@example.com"}}
}
\value{
Slurm job script
}
\description{
A function to list all python scripts to download ERA5 and create a slurm
batch file.
}
\author{
Ahmed Homoudi
}
