% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ERA5_sfc_scripts.R
\name{prepare_ERA5_sfc_scripts}
\alias{prepare_ERA5_sfc_scripts}
\title{Prepare python scripts to get Surface Data from ERA5 Archive}
\usage{
prepare_ERA5_sfc_scripts(
  DATE = "1988-08-5",
  VARS = "34.128/134.128/151.128/165.128/166.128/167.128",
  AREA = "[90, -180, -90, 180,]",
  RES = "0.25/0.25",
  FORMAT = "grib",
  PATH = "~/"
)
}
\arguments{
\item{DATE}{A character indicate on which date (e.g. "1988-08-05") or
time step (e.g. "1988-08-05 13:00:00") the data should be downloaded. default
is "1988-08-05".}

\item{VARS}{A charcter of variables to be downloaded. Please refer to the ERA5 Catalogue.
Default is '34.128/134.128/151.128/165.128/166.128/167.128' which includes:
34 is Sea surface temperature
134 is Surface pressure
151 is Mean sea level pressure
165 is 10 metre U wind component
166 is 10 metre V wind component
167 is 2 metre temperature}

\item{AREA}{A character of the area of interest which is specified as N, W, S, E.
Default is  "square-bracket90, -180, -90, 180,square-bracket".}

\item{RES}{Resolution of the data to be downloaded. Default is "0.25/0.25"}

\item{FORMAT}{Format of the data, either grib or netcdf. Default is grib.}

\item{PATH}{A chacter pointing to where the genereted script will be saved.}
}
\description{
A function to prepare python scripts to download ERA5 data,
mainly for WRF modelling, preferably daily files.
}
\author{
Ahmed Homoudi
}
